+++ tests/integration_tests/test_http_get_ci_schema_v1.py
@@ -5,18 +5,20 @@ import pytest
 from fastapi import status
 
 from app.config import settings
-from app.models.requests import GetCiSchemaV2Params
+from app.models.requests import GetCiSchemaV1Params
+from app.services.ci_classifier_service import CiClassifierService
 from tests.integration_tests.utils import make_iap_request, create_post_params
-from tests.test_config.endpoints import ENDPOINTS, POST_CI, GET_CI_SCHEMA, DELETE_CI
+from tests.test_config.endpoints import ENDPOINTS, GET_CI_SCHEMA_V1, POST_CI, DELETE_CI
 from tests.test_config.endpoints_loader import EndpointsLoader
 
 endpoints_loader = EndpointsLoader(ENDPOINTS)
 
 
-class TestHttpGetCiSchemaV2Restful:
-    """Tests for the `get_collection_instrument_schema_by_guid_v2` endpoint"""
+class TestHttpGetCiSchemaV1:
+    """Tests for the `get_collection_instrument_schema_v1` endpoint"""
 
-    url = endpoints_loader.get_url(GET_CI_SCHEMA)
+    url = endpoints_loader.get_url(GET_CI_SCHEMA_V1)
+
     post_url = endpoints_loader.get_url(POST_CI)
 
     def teardown_method(self):
@@ -30,23 +32,31 @@ class TestHttpGetCiSchemaV2Restful:
     def test_endpoint_returns_400_bad_request_if_bad_query(self):
         """
         What am I testing:
-        `get_collection_instrument_schema_by_guid_v2` should return `HTTP_400_BAD_REQUEST` status if a bad query is made
+        `get_collection_instrument_schema_v1` should return `HTTP_400_BAD_REQUEST` status if a bad query is made
         via a GET request
         """
         # Create a bad querystring
         querystring = urlencode({"my_bad": "querystring"})
-        # Endpoint should return `HTTP_400_BAD_REQUEST
+        # sends request to http_get_ci_schema_v1 endpoint for data
         response = make_iap_request("GET", f"{self.url}?{querystring}")
         assert response.status_code == status.HTTP_400_BAD_REQUEST
 
-    def test_endpoint_returns_404_not_found_if_ci_not_found(self):
+    def test_endpoint_returns_404_not_found_if_ci_not_found(self, setup_payload):
         """
         What am I testing:
-        `get_collection_instrument_schema_by_guid_v2` should return `HTTP_404_NOT_FOUND` status if a valid query is
+        `get_collection_instrument_schema_v1` should return `HTTP_404_NOT_FOUND` status if a valid query is
         made via a GET request but a corresponding ci schema is not found on the db
         """
-        # Create a valid query
-        query_params = GetCiSchemaV2Params(guid="30134e70-c28c-4dcc-b0b0-e403b2df0b24")
+        classifier_type = CiClassifierService.get_classifier_type(setup_payload)
+        classifier_value = CiClassifierService.get_classifier_value(setup_payload, classifier_type)
+
+        # Construct a valid querystring using `setup_payload` data
+        query_params = GetCiSchemaV1Params(
+            classifier_type=classifier_type,
+            classifier_value=classifier_value,
+            language=setup_payload["language"],
+            survey_id=setup_payload["survey_id"],
+        )
         querystring = urlencode(asdict(query_params))
 
         # Endpoint should return `HTTP_404_NOT_FOUND` as no ci exist in the db
@@ -56,35 +66,49 @@ class TestHttpGetCiSchemaV2Restful:
     def test_endpoint_returns_200_success_if_ci_schema_found(self, setup_payload):
         """
         What am I testing:
-        `get_collection_instrument_schema_by_guid_v2` should return `HTTP_200_OK` status if valid ci metadata and schema
+        `get_collection_instrument_schema_v1` should return `HTTP_200_OK` status if valid ci metadata and schema
         exist and a valid query to return the schema is made via a GET request
         """
         # Use `post_ci_v1` to create ci metadata and schema on the db
-
         data = create_post_params(1)
 
         post_response = make_iap_request("POST", f"{self.post_url}?{data[0]}", json=setup_payload)
 
-        created_ci = post_response.json()
-
         assert post_response.status_code == status.HTTP_200_OK
 
-        # Create a valid querystring using the `id` returned when ci created
-        query_params = GetCiSchemaV2Params(guid=created_ci["guid"])
+        classifier_type = CiClassifierService.get_classifier_type(setup_payload)
+        classifier_value = CiClassifierService.get_classifier_value(setup_payload, classifier_type)
+
+        # Construct the querystring to retrieve newly created ci schema
+        query_params = GetCiSchemaV1Params(
+            classifier_type=classifier_type,
+            classifier_value=classifier_value,
+            language=setup_payload["language"],
+            survey_id=setup_payload["survey_id"],
+        )
         querystring = urlencode(asdict(query_params))
-        # sends request to http_get_ci_schema_v2 endpoint for data
+        # sends request to http_get_ci_schema_v1 endpoint for data
         response = make_iap_request("GET", f"{self.url}?{querystring}")
         assert response.status_code == status.HTTP_200_OK
 
-    def test_endpoint_returns_unauthorized_request(self):
+    def test_endpoint_returns_unauthorized_request(self, setup_payload):
         """
         What am I testing:
-        get_collection_instrument_schema_by_guid_v2 should return a 401 unauthorized error if the endpoint is requested with an unauthorized token.
+        get_collection_instrument_schema_v1 should return a 401 unauthorized error if the endpoint is requested with an unauthorized token.
         """
         if settings.CONF == "local-int-tests":
             pytest.skip("Skipping test_endpoint_returns_unauthorized_request on local environment")
 
-        query_params = GetCiSchemaV2Params(guid="30134e70-c28c-4dcc-b0b0-e403b2df0b24")
+        classifier_type = CiClassifierService.get_classifier_type(setup_payload)
+        classifier_value = CiClassifierService.get_classifier_value(setup_payload, classifier_type)
+
+        query_params = GetCiSchemaV1Params(
+            classifier_type=classifier_type,
+            classifier_value=classifier_value,
+            language=setup_payload["language"],
+            survey_id=setup_payload["survey_id"],
+        )
+
         querystring = urlencode(asdict(query_params))
 
         response = make_iap_request("GET", f"{self.url}?{querystring}", unauthenticated=True)
