+++ tests/unit_tests/endpoints/test_http_post_ci_unhappy_paths.py
